fixed routing

+++ src/app/features/breakfast/components/breakfast-card/breakfast-card.component.ts
@@ -3,13 +3,13 @@ import {Breakfast} from "../../models/breakfast-model";
 import {BasketFillingService} from "../../../basket/services/basket-filling.service";
 import {BasketService} from "../../../basket/services/basket.service";
 import {Basket} from "../../../basket/models/basket-model";
-import {HttpClient} from "@angular/common/http";
+import {CurrentUserService} from "../../../../core/services/current-user.service";
+import {map} from "rxjs/operators";
 
 @Component({
   selector: 'app-breakfast-card',
   templateUrl: './breakfast-card.component.html',
-  styleUrls: ['./breakfast-card.component.scss'],
-  providers: [BasketFillingService]
+  styleUrls: ['./breakfast-card.component.scss']
 })
 export class BreakfastCardComponent implements OnInit {
 
@@ -18,50 +18,59 @@ export class BreakfastCardComponent implements OnInit {
 
   inBasket: boolean = false;
 
-  userId: number = 8;
-
   constructor(
     private basketFillingService: BasketFillingService,
-    private basketService: BasketService
-  ) { }
+    private basketService: BasketService,
+    private currentUserService: CurrentUserService
+  ) {
+  }
 
-  ngOnInit(): void {  }
+  ngOnInit(): void {
+  }
 
   handleAddToBasket(): void {
-    this.basketService.getBasketByUserId(this.userId).subscribe(basket => {
-      if (basket == undefined) {
-        const newBasket: Basket = {
-          userID: this.userId,
-          fullPrice: 0,
-          numberOfPersons: 1
-        };
-        this.basketService.createBasket(this.userId, newBasket).subscribe(createdBasket => {
-          this.basketFillingService.addToBasket(createdBasket.basketID, this.breakfast).subscribe(
-            result => {
-              if (result != null) {
-                this.inBasket = true
+    this.currentUserService.user$.pipe(
+      map(user => {
+          if (user.authenticated) {
+            this.basketService.getBasketByUserId(user.info.id).subscribe(basket => {
+              if (basket == undefined) {
+                const newBasket: Basket = {
+                  userID: user.info.id,
+                  fullPrice: 0,
+                  numberOfPersons: 1
+                };
+                this.basketService.createBasket(user.info.id, newBasket).subscribe(createdBasket => {
+                  this.basketFillingService.addToBasket(createdBasket.basketID, this.breakfast).subscribe(result => {
+                    if (result != null) {
+                      this.inBasket = true
+                    }
+                  });
+                })
+              } else {
+                this.basketFillingService.addToBasket(basket.basketID, this.breakfast).subscribe(result => {
+                  if (result != null) {
+                    this.inBasket = true
+                  }
+                });
               }
-            }
-          );
-        })
-      } else {
-        this.basketFillingService.addToBasket(basket.basketID, this.breakfast).subscribe(
-          result => {
-            if (result != null) {
-              this.inBasket = true
-            }
+            });
           }
-        );
-      }
-    });
-
+        }
+      ));
   }
 
   handleRemoveFromBasket(): void {
-    this.basketService.getBasketByUserId(this.userId).subscribe(basket => {
-      this.basketFillingService.removeFromBasket(basket.basketID, this.breakfast.id);
-      this.inBasket = false;
-    });
+    this.currentUserService.user$.pipe(
+      map(user => {
+          if (user.authenticated) {
+            this.basketService.getBasketByUserId(user.info.id).subscribe(basket => {
+              this.basketFillingService.removeFromBasket(basket.basketID, this.breakfast.id);
+              this.inBasket = false;
+            });
+          }
+        }
+      )
+    );
   }
 
   toBase64(img: number[]): string {
